Added support for delayed start of Navigation Co-authored-by: Emre Güder <emrer17m@gmail.com>

+++ App/automotive/src/main/java/com/amos/infotaimos/NavigationPageViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.ViewModel
 
 class NavigationPageViewModel(context: Context) : ViewModel() {
     private var model = Model(context)
-    fun startNavigation(delay: Int) {
+    fun startNavigation(delay: Long) {
         model.startNavigation(delay)
     }
 }
